scripts/requirements.txt
numpy
scipy
soundfile
librosa
openl3
faiss-cpu
tqdm
psycopg2-binary
pgvector
